Refactor: page editor, blocks, page service and repo logic
